rename to threaded_async_lock and at it to hardware_control export
