+++ api/src/opentrons/hardware_control/__init__.py
@@ -21,11 +21,12 @@ from .types import (HardwareAPILike, CriticalPoint,
 from .constants import DROP_TIP_RELEASE_DISTANCE
 from .thread_manager import ThreadManager
 from .execution_manager import ExecutionManager
+from .threaded_async_lock import ThreadedAsyncLock
 
 __all__ = [
     'API', 'Controller', 'Simulator', 'Pipette',
     'SynchronousAdapter', 'HardwareAPILike', 'CriticalPoint',
     'NoTipAttachedError', 'DROP_TIP_RELEASE_DISTANCE',
     'ThreadManager', 'ExecutionManager', 'ExecutionState',
-    'ExecutionCancelledError'
+    'ExecutionCancelledError', 'ThreadedAsyncLock'
 ]
